+++ Tests/ModelsTestHelpers/BucketSplitInfoFixtures.swift
@@ -5,13 +5,10 @@ import ScheduleStrategy
 public final class BucketSplitInfoFixtures {
     public static func bucketSplitInfoFixture(
         numberOfWorkers: UInt = 1,
-        toolResources: ToolResources = ToolResourcesFixtures.fakeToolResources(),
         simulatorSettings: SimulatorSettings = SimulatorSettingsFixtures().simulatorSettings()
-        )
-        -> BucketSplitInfo {
+    ) -> BucketSplitInfo {
         return BucketSplitInfo(
             numberOfWorkers: numberOfWorkers,
-            toolResources: toolResources,
             simulatorSettings: simulatorSettings
         )
     }